Mark data as moved into module
We want to refactor our terraform code to extract some local modules, so that
we can test the modules in isolation.

However, moving the policy into a module changes its address, so by default
terraform will want to recreate the policy the next time we run a plan.  It
looks like there is a way to avoid this by telling terraform where the policy
was moved to.

See https://developer.hashicorp.com/terraform/language/modules/develop/refactoring#splitting-one-module-into-multiple